+++ 2021/day02/lib.h
@@ -6,8 +6,8 @@
 
 namespace day02
 {
-    size_t run_part1(const std::vector<std::string> &input);
-    size_t run_part2(const std::vector<std::string> &input);
+    int32_t run_part1(const std::vector<std::string> &input);
+    int32_t run_part2(const std::vector<std::string> &input);
 }
 
 #endif
