+++ 2021/day04/lib.h
@@ -6,8 +6,8 @@
 
 namespace day04
 {
-    size_t run_part1(const std::vector<std::string> &input);
-    size_t run_part2(const std::vector<std::string> &input);
+    uint32_t run_part1(const std::vector<std::string> &input);
+    uint32_t run_part2(const std::vector<std::string> &input);
 }
 
 #endif
